feat: [AD-008] extracted sidebar to seperate file
