+++ view/src/App.tsx
@@ -1,11 +1,9 @@
-import { useState, useEffect } from "react";
-import { TombacApp, Button, DatePicker } from "tombac";
-import AnomalyList from "./components/AnomalyList";
+import { useState } from "react";
+import { TombacApp } from "tombac";
 import MapView from "./components/MapView";
 import BottomBar from "./components/BottomBar";
+import Sidebar from "./components/Sidebar";
 import { AppProvider, useAppContext } from "./AppContext";
-import { fetchAnomalyData } from "./api/api";
-import { CSSProperties } from "react";
 
 const containerStyle = {
   display: "flex",
@@ -15,66 +13,12 @@ const containerStyle = {
   padding: 0,
 };
 
-const sidebarStyle: CSSProperties = {
-  width: 280,
-  borderRight: "1px solid #ccc",
-  padding: 16,
-  overflowY: "auto",
-};
-
-function addMonths(date: Date, months: number) {
-  const d = new Date(date);
-  d.setMonth(d.getMonth() + months);
-  return d;
-}
-
-function subtractMonths(date: Date, months: number) {
-  const d = new Date(date);
-  d.setMonth(d.getMonth() - months);
-  return d;
-}
-
 function AppContent() {
-  const {
-    mode,
-    setMode,
-    anomalyGeoJson,
-    filteredFeatures,
-  } = useAppContext();
+  const { mode, anomalyGeoJson, filteredFeatures } = useAppContext();
 
-  const [startDay, setStartDay] = useState<Date>(new Date("2025-01-01T00:00"));
-  const [endDay, setEndDay] = useState<Date>(new Date("2025-02-01T00:00"));
   const [selectedPolygon, setSelectedPolygon] = useState<any>(null);
 
-  useEffect(() => {
-    const minEnd = addMonths(startDay, 1);
-    if (endDay < minEnd) setEndDay(minEnd);
-  }, [startDay]);
-
-  useEffect(() => {
-    const minStart = subtractMonths(endDay, 1);
-    if (startDay > minStart) setStartDay(minStart);
-  }, [endDay]);
-
   if (!anomalyGeoJson) return <div>Loading anomalies...</div>;
-
-  const handleGenerateReport = async () => {
-    if (!selectedPolygon) {
-      alert("Please select a polygon to generate the report.");
-      return;
-    }
-    try {
-      await fetchAnomalyData({
-        startDay: startDay.toISOString().slice(0, 19),
-        endDay: endDay.toISOString().slice(0, 19),
-        coordinates: selectedPolygon.geometry.coordinates[0],
-        dataType: "TOTAL_DISTANCE_M",
-      });
-      setMode("viewing");
-    } catch (error: any) {
-      alert("Failed to generate report: " + error.message);
-    }
-  };
 
   return (
     <TombacApp
@@ -82,45 +26,7 @@ function AppContent() {
       theme={{ baseUnit: "px", settings: { modalZIndex: 20 } }}
     >
       <div style={containerStyle}>
-        {(mode === "viewing" || mode === "drawing") && (
-          <aside style={sidebarStyle}>
-            {mode === "drawing" && (
-              <>
-                <DatePicker
-                  value={startDay}
-                  onChange={(date) => {
-                    if (date instanceof Date) setStartDay(date);
-                  }}
-                  maxDate={subtractMonths(endDay, 0)}
-                />
-                <DatePicker
-                  value={endDay}
-                  onChange={(date) => {
-                    if (date instanceof Date) setEndDay(date);
-                  }}
-                  minDate={addMonths(startDay, 1)}
-                />
-                <Button
-                  onClick={handleGenerateReport}
-                  style={{ marginTop: 16 }}
-                >
-                  Generate Report
-                </Button>
-              </>
-            )}
-            {mode === "viewing" && (
-              <>
-                <AnomalyList />
-                <Button
-                  onClick={() => setMode("drawing")}
-                  style={{ marginTop: 16 }}
-                >
-                  Back to Drawing
-                </Button>
-              </>
-            )}
-          </aside>
-        )}
+        <Sidebar selectedPolygon={selectedPolygon} />
 
         <main style={{ flex: 1, position: "relative" }}>
           <MapView
